requirements.txt
amqp==5.0.6
arrow==1.1.1
attrdict==2.0.1
billiard==3.6.4.0
blinker==1.4
boto3==1.18.3
botocore==1.21.3
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
configparser==5.0.2
cryptography==3.4.7
dataclasses==0.6
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Mail==0.9.1
Flask-Minio==0.1.2
Flask-WTF==0.15.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
kombu==5.1.0
lxml==4.6.3
MarkupSafe==2.0.1
minio==6.0.2
msal==1.12.0
ntlm-auth==1.5.0
Office365-REST-Python-Client==2.3.4
prompt-toolkit==3.0.19
pycparser==2.20
PyJWT==2.1.0
pyminio==0.2.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.18.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
s3transfer==0.5.0
SharePlum==0.5.1
six==1.16.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3
